Fix: high dbus usage.
The module always refreshed itself, if the AC adapter was read. This seemed to
cause a dmus-message-storm, which caused other programs to stop function or be
very slow. This is fixed now. The update function (power_supply_changed()) is
only called, if there is a change. The test, if there was a change, is put into
the callback for the properties.

I still don't understand, why the call to power_supply_changed() is necessary
at all.